chore: fix all clippy warnings
- Use let-chain syntax for nested if-let patterns (Rust 2024)
- Use div_ceil(), is_multiple_of(), is_none_or() methods
- Use contains() for slice membership checks
- Use unsigned_abs() for safe absolute value conversion
- Add is_empty() to EvidenceSource trait
- Add #[allow(clippy::too_many_arguments)] for pipeline functions
- Remove redundant borrows, casts, and bindings
- Add truncate(true) to probe file creation
- Use idiomatic patterns: strip_prefix, eq_ignore_ascii_case, Range::contains

+++ src/carve/rar.rs
@@ -256,15 +256,14 @@ fn parse_rar5(
             return Err(CarveError::Invalid("rar5 header size invalid".to_string()));
         }
 
-        let header_buf =
-            match read_exact_at(ctx, offset + 4 + size_len as u64, header_size as usize) {
-                Some(buf) => buf,
-                None => {
-                    truncated = true;
-                    errors.push("eof while reading RAR header".to_string());
-                    break;
-                }
-            };
+        let header_buf = match read_exact_at(ctx, offset + 4 + size_len, header_size as usize) {
+            Some(buf) => buf,
+            None => {
+                truncated = true;
+                errors.push("eof while reading RAR header".to_string());
+                break;
+            }
+        };
 
         let mut idx = 0usize;
         let header_type = read_varint_buf(&header_buf, &mut idx)
